refactor: split SQLite metadata modules
