+++ .github/workflows/ci.yml
@@ -189,7 +189,7 @@ jobs:
           export PATH="$HOME/.local/sqlite-below-minimum/bin:$PATH"
           sqlite3 --version
           SABIQL_EXPECT_SQLITE_SAFE_MODE_REJECTION=1 cargo test -p sabiql-infra \
-            adapters::sqlite::sqlite3::metadata::tests::metadata::rejects_sqlite_before_safe_mode_minimum_at_connection \
+            adapters::sqlite::metadata::catalog::tests::metadata::rejects_sqlite_before_safe_mode_minimum_at_connection \
             -- --exact
           cargo test -p sabiql-app \
             model::connection::error::tests::error_info::from_db_operation_error_classifies_sqlite_safe_mode_requirement \
