+++ src/infra/adapters/sqlite/sqlite3/parser/mod.rs
@@ -1,5 +1,5 @@
 mod command_tag;
-pub(in crate::adapters::sqlite::sqlite3) mod lexer;
+mod lexer;
 mod output;
 
 pub(super) use command_tag::{
@@ -10,6 +10,9 @@ pub(super) use lexer::{
     SqliteStatementPlan, append_changes_query_for_plan, is_sqlite_rerunnable_export_query,
     reject_sqlite_fsdir, sqlite_adhoc_execution_query_for_plan, sqlite_empty_result_sentinel,
     sqlite_export_not_rerunnable_error, sqlite_probe_marker, sqlite_statement_plan,
+};
+pub(in crate::adapters::sqlite) use lexer::{
+    is_create_view_prefix, is_create_virtual_table_prefix, virtual_table_module_name,
 };
 pub(super) use output::{
     last_sqlite_result_set, parse_affected_rows, parse_count_result, quoted_to_query_result,
